Form Styles and User Interaction
I continued with form styles and tweaks to the blur and focus inter-
actions. I created a computed function to handle the logic for
swapping classes when user focuses and blurs to work in conjunction
the method that handles this event. I figured it would be a better
practice to remove the logic from the html.