Media Inventory List
Created a new way to display and style the media inventory list, its
based on a rough design I did in XD. I also created and added an edit
and delete icon. I did a lot of styling and reset of specific elements,
I know I can pull-in normalize.css but I'm trying to keep this project
as slim as possible. I'm leaving the styles I added in my attempt to
make the footer sticky. I will address later.