Logic for Skeleton Screens
Createed a bit of logic for the skeleton screens component I added to
app.vue. I did encounter a linting error when I used v-for and v-if
together. Apparently, Vue.js deems the use of these two directives as
an improper so I mickey moused it by moving the v-if to a presentational
div. I'll be coming back to this later as I need to fully understand the
why. For the time being the solution I implemented works and doesn't throw
any errors.