Stat Component
Created a stat component so I can keep track of how many types of
media I have in the database. This is based off the original code I
used to get the total from the database. It still needs tweaking but
at least I can see what I have and match it to my Plex server.