Stats and Donuts
Refactored the Stat component html, scripts and styles. The refactor
simplified the component by moving the v-for out of it and into the
parent instead. I also fixed the computed method I used to filter
the db object I'm passing a long with naming conventions used so they
don't collide with native ones.