Inventory and Actions
Refactored the inventory code that was hanging out in App.vue and
moved it into its own vue component along with the action buttons. I
also modified the modwrap component to dynamically set modifier classes
and changed the --grid modifier code to be a responsive grid. I'll have
to apply the same to the skeleton.