Login and InputField
Continued work on the inputfield component and login view. The bulk of
the work went into getting these two components to communicate with
each other. The inputfield focus/events still need the logic to be
tweaked so the label isn't animated when there is a value in the field.

There are also various styles that were added to add a bit of
interactive feedback for the user.