Dashboard + NewMedia Components
Moved all the Dashboard code from App.vue to its own custom component
so I can keep App.vue as clean as possible.

I also updated the NewMedia.vue component and expanded some of the
fields including collection so it displays a input text if the user
select yes. I'll need to sync with the Lambda function so I can take
the object I'll be creating when there is a collection.

Finally, I applied some styles to Login.vue.