Notifying Grand Pappy!
Applied some voodoo to Actions.vue to capture the selected inventory
items id and emit the value to the grandparent component. I also
updated Dashboard.vue to include the inventory id and added the
click event to receive and invoke deleteMedia function. Further work
is needed to actually get it to delete the entry in DB.