Sheet Component
Created a new component called Sheet. This component will be used
to add new inventory items. The idea is to more the NewMedia component
into it. This component is in its infancy with only a bit of logic and
styles. I also added a new add icon for the button in the component and
updated Dashboard component to include a window.confirm() just stop
accidental deletion of items during development. I'll later refine this
when I create my own Dialog component.