Format Modifiers and Computed Classes
Created a new computed function to set a class based on the format
received from the API. I also created the styles to give each modifier
a color and text based off it. I changed the dimensions of the poster
a bit from the design and updated the placeholder image to reflect the
change.