Inventory Data, Loading, and Getters
Created "getters" to aid in my quest of integrating the store further
into my existing components. Two of the "getters" are for inventory and
the other for the loader.

I updated the dashboard component to use the "actions" and "getters" I
created in the store. At this point the API data is displayed in the
dashboard again but I will be moving this to its own page after I
update the rest of the components.