Dashboard Styles
I updated the dashboard styles to follow the design I created. This is
still a wip. For the moment, the doughnut chart is consistently being
displayed without any of the random disappearances I encountered in the
previous iteration. I also changed the background color I was using in
the dashboard as it was too light. In order to override the slot styles
I created a new class `dashboard-pg`. Until, I understand/upgrade to
latest VueJS this is what I'll be adding to new views. The whole
slot/scoped/slotted() was not working at this time.