Donut Chart Render Issues and Missing Data
Fixed the donut chart not rendering because of the dataset being empty.
The issue was happening because I was creating the donut data in the
`mounted()` hook without waiting for the `getInventory()` to wrap up.
To resolve the issue I added async to the `mounted()` hook and wait for
the data to be properly set.

I created additional methods in the store to count the inventory format
so I can set the donut segments. Eventually, I'll use this for the
other stats I'm displaying. Finally, I cleaned up the commented code in
the doughnutchart component.