Login, Authentication, AccountMenu and Button
Continued development of the login and authentication logic. Created a
`login()`, `logout()` actions in the store and moved code I initially
created for login to it. I commented the safeguard code I added to
routes because it keeps throwing an error about using `next()` more
than once. I'll issue a fix once I get the login component done.

I created two new components: accountmenu and basebutton. These were
added and replaced html I was using in login and header. The
accountmenu component was necessary to add the logout functionality.
The basebutton was created to be a reusable component.