Dashboard Redesign (WIP)
Began applying the design on the dashboard component. I created a new
ChartStat component to help create the legend and display total counts.
This is similar to the Stat component I created earlier in the project
but without the overall complexity. I used scoped and non-scoped styles
so I can override them in parent components, I'm not sure if this is
the best solution but I will research it more at a later time.

I forgot to include the Brand component I created in a previous commit
so I've added tot his one. Nothing crazy, just made the logo into a
component so I can make changes and tweaks to it in one place.