Dashboard Store Updates
Changed the name of `calcDonutSegments` action to `calDonutLegend`
because I refactored it to be a bit more versatile in tabulating the
filtered data count. I also figured out how to send multiple parameters
to actions when using `dispatch()`. Lastly, I updated the Dashboard
component to update with the new numbers.