Search Component w/Delay (It's Alive!)
Finished working on the Search component functionality. I ended using
an emit to pass the filtered data to the parent. I'll have to look into
using it in computed within the child. I also updated the `setTimeout()`
logic and got it working by using `this` since a it references itself
rather than what you would expect normally. I left a comment explaining
it for future generation.

In conjunction with the Search component I made changes to the
Inventory/Details component to reflect changes required for the data to
be used after its been filtered. I also updated the details prop to
match the props I'm getting from the API.