Fetching Inventory
Issued a fix on the inventory component not loading inventory items on
refresh by updating `created()` lifecycle hook with `async` and waiting
for the `this.getInventory()` to wait for the data. I applied the same
to Dashboard `mounted()` and removed `created()`. The reason this is
placed in the `mounted()` is because of the way the chart library is
invoked. I also removed the `try/catch` as it was redundant because the
`getInventory()` already has it.