The Big Bang
I upgraded from Vue 3 to 4 and on the latest version of NodeJS, version
v16.18.0. The upgrade was not straightforward, I keep getting errors
using `vue upgrade`. This was done on a clean NodeJS installation. I
resolved it by creating a fake project using the CLI and appending the
changes and removing files. Afterwards, I had to make additional
changes to components pointing to the wrong image folder. The upgrade
has an 'asset' folder in 'src' now. Here are other changes I made:

- Moved 'public/images' to 'src/assets'
- Removed `tag` attribute from `router-link` as its deprecated
- Updated vue.config.js to match upgrade
- Deleted .browserslistcrc as its now included in package.json
- Created jsconfig.json to match upgrade
- Updated package.json to match upgrade but added extra props to better
  define it