Create Component and Adding New Items
Bulk of development went into getting the base functionality of the
form working enough to insert new items into the database. I added a
few `refs` so I can access InputFile method and track of the values set
for the InputRadio's.

I also created a new action in the store to add items but I still need
to update the logic in my lambda function to return the item back so I
can update the store without getting all the inventory data again.
Finally, here's a list of other things I did:

- Updated InputRadio to set ModelValue on focus set by `mounted()`
- Moved NewMedia component to archive and removed references
- Removed unnecessary variables from App.vue
- Disabled up/down arrows from input fields with type text
- Updated InputText to bind $attrs and handle type text fields