New GetInventory Lambda, Archived Components and Lambda Functions
I refactored the previous Lambda function I was using to get all the
inventory items. I stored the previous in the archived folder and gave
its own folder along with components. I know I should just delete them
but I might have to reference them in the future. Once the app is
deployed to prod I'll delete and cleanup app leavings.

I accidentally included an InputTextArea components I was working on
and rather than resetting the HEAD and doing all kinds of Git magic
I decided to just leave it in this commit.