Lambda New-Item Overhaul
I did a complete overhaul on the 'new-item' lambda function. This was a
necessary evil because I needed to parse multipart/form-data and upload
an image file to S3. During development index.js became a maintenance
mess so I decided to split into three modules. These functions are in
working order and fully tested in Lambda.

There is one minor hitch, the
file upload cannot be read by Windows Explorer but shows up fine in
Photoshop. I'm certain its an issue on my end, I'll issue a fix once I
clean up the rest of the Lambda functions and consolidate some of the
console.logs.