Convert Image Lambda Function and README
Created 'convert-image' lambda function to convert the images I upload
to 'noface-shelfdat.com' bucket. The function is in working order but I
need to go back to the 'new-item' lambda function to correctly buffer
the image as it throws errors unless I directly upload an image.

I also updated the README file and added a note about an alternative
method for installing and using Sharp libraries in serverless
environment.