Store Changes For AddNewItem Function
Made a few tweaks to `addNewItem()` to get the multipart/form-data to
work correctly. Because we are using multipart/form-data I no longer
needed to stringify the payload or setting content-type. Also,
content-type for multipart/form-data was not necessary as it was
setting boundary incorrectly and making the parser thrown a bunch of
errors.