Separating The Year
I issued a fix for the year being added to the subtitle. The cause was
the pattern and map() logic I was using to break apart the title data
I'm getting from the server. It worked as long as I had all three
elements. But choked when there wasn't a subtitle. I updated
'putter.js' to insert the year into the database, added the prop to
Details component, and Inventory view.