Jenkins Build Errors
Issued a temporary fix to get Jenkins building and deploying. The issue
is caused because the '.certs' folder is not in the project when it
pulls the repo from GitLab. This is by design as that folder should be
in the repo and is meant for local development. I mickey moused a fix
but I left a todo to come back and issue a real fix.