Sticky Footer Not Sticky
Issued a fix to the footer not sticking to bottom of visible viewport
on mobile/tablet devices. At first, the issue is not apparent and does
not appear in local debugging with dev console switched to responsive
mode.

I ended up using a mobile/tablet detection library as there have been
reports of latest versions of iOS reporting as MacOS. Plus, I feel a
library is the better choice versus creating my own.