NewItem Lambda and Anchor Route Trigger
Issued a fix for the new-item lambda sending integers as strings. The
components props for details expects an integer which is correctly set
in the DB but not when sending the new-item data.

I also fixed the poster component triggering a route change because the
image was wrapped in an anchor. The issue bubbled up because I had set
logic in the routes to scroll the content up on route change.